Milestone4/requirements.txt
streamlit
pandas
plotly
fpdf
PyPDF2
python-docx
